+++ src/core_plugins/kibana/public/kibana.js
@@ -59,7 +59,6 @@ import 'ui/agg_types';
 import 'ui/timepicker';
 import { showAppRedirectNotification } from 'ui/notify';
 import 'leaflet';
-import { KibanaRootController } from './kibana_root_controller';
 
 routes.enable();
 
@@ -67,7 +66,5 @@ routes
   .otherwise({
     redirectTo: `/${chrome.getInjected('kbnDefaultAppId', 'discover')}`
   });
-
-chrome.setRootController('kibana', KibanaRootController);
 
 uiModules.get('kibana').run(showAppRedirectNotification);
