--- src/core_plugins/kibana/public/kibana_root_controller.js
@@ -1,34 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import moment from 'moment-timezone';
-
-export function KibanaRootController($scope, courier, config) {
-  config.watch('dateFormat:tz', setDefaultTimezone, $scope);
-  config.watch('dateFormat:dow', setStartDayOfWeek, $scope);
-
-  function setDefaultTimezone(tz) {
-    moment.tz.setDefault(tz);
-  }
-
-  function setStartDayOfWeek(day) {
-    const dow = moment.weekdays().indexOf(day);
-    moment.updateLocale(moment.locale(), { week: { dow } });
-  }
-}
