+++ src/ui/public/autoload/all.js
@@ -21,4 +21,5 @@ import './accessibility';
 import './modules';
 import './directives';
 import './filters';
+import './settings';
 import './styles';
